Add 'flatten_single' reduce function
This reduce function expect the input `values` list contains only one
value, and returns it. It throws an exception if it is not the case.
This reduce function is useful when a pattern should appear only once
in the log file preventing to report wrong output when the input file
contains multiple instances of the context. This happens particularly
for the log files when the program appends to the existing log file
instead of rewriting.